requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.109.1
fastapi-utils==0.2.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
iso8601==1.1.0
pydantic==1.10.14
pydantic_core==2.3.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
tortoise-orm==0.20.0
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
